More abstract typed arguments, moved the defaults into their own package
