+++ src/main/java/com/njdaeger/pdk/command/brigadier/arguments/AbstractStringTypedArgument.java
@@ -1,0 +1,13 @@
+package com.njdaeger.pdk.command.brigadier.arguments;
+
+import com.mojang.brigadier.arguments.ArgumentType;
+import com.mojang.brigadier.arguments.StringArgumentType;
+import org.jetbrains.annotations.NotNull;
+
+public abstract class AbstractStringTypedArgument<TYPE> extends BasePdkArgumentType<TYPE, String> {
+
+    @Override
+    public @NotNull ArgumentType<String> getNativeType() {
+        return StringArgumentType.word();
+    }
+}
